Update FakeController and its Config for testing
* Pulled `FakeController` and `FakeControllerConfig` out of ControllersManagerSpec.
* Replaced `FakeControllerManager.Config` with `FakeControllerConfig`.
* Moved `FakeMappingManager` and `FakeControllerManager` to separate files.
